Extract TrackContainer to a separate file
